enterprise/RemoveShixin/requirements.txt
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
cryptography==2.4.2
docx==0.2.4
et-xmlfile==1.0.1
fake-useragent==0.1.11
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
jdcal==1.4
lxml==4.2.5
numpy==1.15.4
openpyxl==2.5.12
pandas==0.23.4
pd==0.0.1
Pillow==5.3.0
psycopg2==2.7.6.1
pycparser==2.19
PyExecJS==1.5.1
PyHamcrest==1.9.0
PyMySQL==0.9.3
python-dateutil==2.7.5
python-docx==0.8.7
pytz==2018.7
redis==3.0.1
requests==2.20.1
selenium==3.141.0
six==1.11.0
urllib3==1.24.1
xlrd==1.1.0
zope.interface==4.6.0
